requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.3
idna==3.4
lxml==4.9.3
psycopg2-binary==2.9.6
requests==2.31.0
sqlparse==0.4.4
urllib3==2.0.4
